+++ include/amath/vec4.h
@@ -124,8 +124,8 @@ struct Vec4 {
    }
 
    // Access
-   scalar operator[](const int i) const { return d[i]; }
-   scalar &operator[](const int i) { return d[i]; }
+   scalar operator[](const size_t i) const { return d[i]; }
+   scalar &operator[](const size_t i) { return d[i]; }
 };
 
 }  // namespace amath
